Replace ContentEncodingTrait to TagContentTrait
